# docs/swagger.yaml
definitions:
...
  dto.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      kind:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /car-management/create:
    post:
      consumes:
      - application/json
      description: A POST Request to create a car and assigned to the driver whose
        creating it
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: |-
          FrontCarPhoto     *multipart.FileHeader `json:"frontCarPhoto" swaggerignore:"true"`
          LicensePhoto      *multipart.FileHeader `json:"licensePhoto" swaggerignore:"true"`
          RoadTaxPhoto      *multipart.FileHeader `json:"roadTaxPhoto" swaggerignore:"true"`
          VEPPhoto          *multipart.FileHeader `json:"vepPhoto" swaggerignore:"true"`
        in: formData
        name: carManagementType
        required: true
        type: string
      - in: formData
        name: carModelId
        required: true
        type: string
      - in: formData
        name: name
        required: true
        type: string
      - in: formData
        name: plateNumber
        required: true
        type: string
      - description: front car photo upload
        in: formData
        name: frontCarPhoto
        required: true
        type: file
      - description: license photo upload
        in: formData
        name: licensePhoto
        required: true
        type: file
      - description: road tax photo upload
        in: formData
        name: roadTaxPhoto
        required: true
        type: file
      - description: vep photo upload
        in: formData
        name: vepPhoto
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to create or assign a car to the driver
      tags:
      - Car Management
  /car-management/delete/{id}:
    delete:
      consumes:
      - application/json
      description: A DELETE Request to delete record car management
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to delete car management made by driver
      tags:
      - Car Management
  /car-management/get:
    get:
      consumes:
      - application/json
      description: A GET Request to fetch a all records for driver to view list of
        added car management that already created
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to get all car management made by driver
      tags:
      - Car Management
  /car-management/set/active/{id}:
    put:
      consumes:
      - application/json
      description: A UPDATE Request to set record car management to active
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to set active car management made by driver
      tags:
      - Car Management
  /car-management/set/inactive/{id}:
    put:
      consumes:
      - application/json
      description: A UPDATE Request to set record car management to inactive
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to set inactive car management made by driver
      tags:
      - Car Management
  /car-management/update/{id}:
    put:
      consumes:
      - application/json
      description: A UPDATE Request to update record car management
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: carModelId
        required: true
        type: string
      - in: formData
        name: licensePhoto
        type: string
      - in: formData
        name: name
        required: true
        type: string
      - in: formData
        name: plateNumber
        required: true
        type: string
      - description: front car photo upload
        in: formData
        name: frontCarPhoto
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to update car management made by driver
      tags:
      - Car Management
  /company/create:
    post:
      consumes:
      - application/json
      description: A POST Request to create a company
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: companyAddress
        required: true
        type: string
      - in: formData
        name: companyCountry
        required: true
        type: string
      - in: formData
        name: companyName
        required: true
        type: string
      - in: formData
        name: companyNumber
        required: true
        type: string
      - in: formData
        name: email
        required: true
        type: string
      - in: formData
        name: name
        required: true
        type: string
      - in: formData
        name: phoneNumber
        required: true
        type: string
      - description: company certificate upload
        in: formData
        name: companyCertificate
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to create company
      tags:
      - Company
  /company/create/approval/{id}:
    get:
      consumes:
      - application/json
      description: A POST Request to accept company submission request
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to Approve company registration
      tags:
      - Company
  /company/driver/all/{id}:
    get:
      consumes:
      - application/json
      description: A GET Request to fetch a all records
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to get all drivers by company id
      tags:
      - Company
...
swagger: "2.0"